requirements/base.txt
Django==1.6.5
bpython==0.13
django-braces==1.4.0
django-model-utils==2.0.3
django-pure-pagination==0.2.1
Pygments==1.6
logutils==0.3.3
South==0.8.4
django-extensions
django-debug-toolbar
